# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ระบบทดสอบ API ด้วย Fiber v3",
        "title": "Fiber Test API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/names": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูล Names ทั้งหมดในระบบ (ต้อง Login)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "names"
                ],
                "summary": "ดึงรายชื่อ (Get Names)",
                "responses": {
                    "200": {
                        "description": "คืนค่าเป็น Array ของ Name",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Name"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/names/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูล Name ตาม ID ที่ระบุ (ต้อง Login)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "names"
                ],
                "summary": "ดึงรายชื่อตาม ID (Get Name By Id)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Name ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Name"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "ส่ง Username/Password เพื่อรับ JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "ข้อมูล Login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Name": {
            "type": "object",
            "properties": {
                "fname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lname": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "พิมพ์คำว่า \"Bearer \" ตามด้วย Token (เช่น \"Bearer eyJhb...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}